+++ packages/terra-content-container/tests/wdio/content-container-spec.js
@@ -1,12 +1,12 @@
 Terra.describeViewports('Content Container', ['medium'], () => {
   it('displays default Content Container', () => {
-    browser.url('/raw/tests/terra-content-container/content-container/default-content-container');
+    browser.url('/raw/tests/cerner-terra-core-docs/content-container/default-content-container');
 
     Terra.validates.element('content container');
   });
 
   it('displays Content Container with fill', () => {
-    browser.url('/raw/tests/terra-content-container/content-container/fill-content-container');
+    browser.url('/raw/tests/cerner-terra-core-docs/content-container/fill-content-container');
 
     Terra.validates.element('content container fill');
   });