+++ packages/terra-demographics-banner/src/terra-dev-site/test/demographics-banner/GestationalAgeDemographicsBannerNoLabel.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import DemographicsBanner from '../../../DemographicsBanner';
-
-export default () => <DemographicsBanner id="gestational-age-no-label" gestationalAge="April 5, 2016" />;