+++ packages/terra-card/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
+* Changed
+  * Migrate doc to terra-core-docs.
+
 ## Unreleased
+
+## 3.42.0 - (November 2, 2021)
+
+* Changed
+  * Minor dependency version bump
 
 ## 3.41.0 - (September 21, 2021)